requirements.txt
Flask==3.0.3
gunicorn>=21.2
pymongo[srv]==4.7.2
nltk==3.8.1
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
llama-cpp-python==0.2.54         # CPU‑only wheel, includes BLAS
